refactor(ui-grid): extract stored state
